# src/main/webapp/docs/swagger.yaml
swagger: "2.0"
info:
  description: "An anime api where you can get anime photos, videos, news and merchandising. Also you can update your MyAnimeList account."
  version: 1.0.0
  title: "Proyecto-Miyamoto"
  # put the contact info for your development or API team
  contact:
    email: "ernesto.saquete.rios@gmail.com"

  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"

# all paths will be relative to the basePath specified
# basePath: /v1
# Added by API Auto Mocking Plugin
# host: proyectomiyamoto.appspot.com/
# basePath: /api
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /ernsaqrio/Proyecto-Miyamoto/1.0.0
...
schemes:
  - "https"
paths:
  "/noticias":
    get:
      tags:
      - "Recurso Noticia"
      summary: "Método que devuelve todas las noticias"
      operationId: "getAllNoticias"
      produces:
        - "application/json"
      parameters: []
      responses:
        200:
          description: "Listado de todas las noticias"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Noticia"
        404:
          description: "Ningún resultado"
  "/noticias/idAnime":
    get:
      tags:
      - "Recurso Noticia"
      summary: "Método que devuelve todas las noticias de la id de un determinado anime"
      operationId: "getAllNoticiasById"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "id"
          description: "ID del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todas las noticias de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Noticia"
        404:
          description: "Ningún resultado"
  "/noticias/titleAnime":
    get:
      tags:
      - "Recurso Noticia"
      summary: "Método que devuelve todas las noticias del nombre de un determinado anime"
      operationId: "getAllNoticiasByTitle"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "title"
          description: "Título del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todas las noticias de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Noticia"
        404:
          description: "Ningún resultado"
  "/noticias/{id}":
    get:
      tags:
      - "Recurso Noticia"
      summary: "Método que devuelve una noticia con un determinado id"
      operationId: "getNoticiaById"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID de la noticia"
          required: true
          type: "string"
      responses:
        200:
          description: "Noticia según un determinado id"
          schema:
            $ref: "#/definitions/Noticia"
        404:
          description: "Ningún resultado"
  "/imagenes":
    get:
      tags:
      - "Recurso Imagen"
      summary: "Método que devuelve todas las imágenes"
      operationId: "getAllImagenes"
      produces:
        - "application/json"
      parameters: []
      responses:
        200:
          description: "Listado de todas las imágenes"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Imagen"
        404:
          description: "Ningún resultado"
  "/imagenes/idAnime":
    get:
      tags:
      - "Recurso Imagen"
      summary: "Método que devuelve todas las imágenes de la id de un determinado anime"
      operationId: "getAllImagenesById"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "id"
          description: "ID del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todas las imágenes de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Imagen"
        404:
          description: "Ningún resultado"
  "/imagenes/titleAnime":
    get:
      tags:
      - "Recurso Imagen"
      summary: "Método que devuelve todas las imagenes del nombre de un determinado anime"
      operationId: "getAllImagenesByTitle"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "title"
          description: "Título del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todas las imágenes de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Imagen"
        404:
          description: "Ningún resultado"
  "/imagenes/{id}":
    get:
      tags:
      - "Recurso Imagen"
      summary: "Método que devuelve una noticia con un determinado id"
      operationId: "getImagenById"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID de la imagen"
          required: true
          type: "string"
      responses:
        200:
          description: "Imagen según un determinado id"
          schema:
            $ref: "#/definitions/Imagen"
        404:
          description: "Ningún resultado"
  "/videos":
    get:
      tags:
      - "Recurso Video"
      summary: "Método que devuelve todos los vídeos"
      operationId: "getAllVideos"
      produces:
        - "application/json"
      parameters: []
      responses:
        200:
          description: "Listado de todos los vídeos"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Video"
        404:
          description: "Ningún resultado"
  "/videos/idAnime":
    get:
      tags:
      - "Recurso Video"
      summary: "Método que devuelve todos los vídeos de la id de un determinado anime"
      operationId: "getAllVideosById"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "id"
          description: "ID del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todos los vídeos de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Video"
        404:
          description: "Ningún resultado"
  "/videos/titleAnime":
    get:
      tags:
      - "Recurso Video"
      summary: "Método que devuelve todos los vídeos del nombre de un determinado anime"
      operationId: "getAllVideosByTitle"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "title"
          description: "Título del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todos los vídeos de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Video"
        404:
          description: "Ningún resultado"
  "/videos/{id}":
    get:
      tags:
      - "Recurso Video"
      summary: "Método que devuelve un vídeo con un determinado id"
      operationId: "getVideoById"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID del vídeo"
          required: true
          type: "string"
      responses:
        200:
          description: "Vídeo según un determinado id"
          schema:
            $ref: "#/definitions/Video"
        404:
          description: "Ningún resultado"
  "/productos":
    get:
      tags:
      - "Recurso Producto"
      summary: "Método que devuelve todos los productos"
      operationId: "getAllProductos"
      produces:
        - "application/json"
      parameters: []
      responses:
        200:
          description: "Listado de todos los productos"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Producto"
        404:
          description: "Ningún resultado"
  "/productos/idAnime":
    get:
      tags:
      - "Recurso Producto"
      summary: "Método que devuelve todos los productos de la id de un determinado anime"
      operationId: "getAllProductosById"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "id"
          description: "ID del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todos los productos de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Producto"
        404:
          description: "Ningún resultado"
  "/productos/titleAnime":
    get:
      tags:
      - "Recurso Producto"
      summary: "Método que devuelve todos los productos del nombre de un determinado anime"
      operationId: "getAllProductosByTitle"
      produces:
        - "application/json"
      parameters:
        - in: "query"
          name: "title"
          description: "Título del anime"
          required: true
          type: "string"
      responses:
        200:
          description: "Listado de todos los productos de un determinado anime"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Producto"
        404:
          description: "Ningún resultado"
  "/productos/{id}":
    get:
      tags:
      - "Recurso Producto"
      summary: "Método que devuelve un vídeo con un determinado id"
      operationId: "getProductoById"
      produces:
        - "application/json"
      parameters:
        - in: "path"
          name: "id"
          description: "ID del vídeo"
          required: true
          type: "string"
      responses:
        200:
          description: "Vídeo según un determinado id"
          schema:
            $ref: "#/definitions/Producto"
        404:
          description: "Ningún resultado" 
...
definitions:
...
  Imagen:
    required:
      - id
      - size
      - url
      - title
    properties:
      id:
        type: string
        example: i1
      size:
        type: string
        example: 1080x920
      url:
        type: string
        example: https://68.media.tumblr.com/c705d57e15bb02a9bbcfe2f49e2267be/tumblr_oq2vo67Zm61t6nfoqo1_1280.jpg
      title:
        type: string
        example: Cosplay higurashi
  Noticia:
    required:
      - id
      - title
      - url
      - content
    properties:
      id:
        type: string
        example: n1
      title:
        type: string
        example: New promo
      url:
        type: string
        example: http://www.animenewsnetwork.com/news/2011-07-21/higurashi-no-naku-koro-ni-kira-magical-girl-promo-posted
      content:
        type: string
        example: Animate began streaming a "Magical Girl" promotional video for the Higurashi no Naku Koro ni Kira original video anime (OVA) series on Thursday.
  Producto:
    required:
      - id
      - name
      - url
      - prize
      - description
    properties:
      id:
        type: string
        example: p1
      name:
        type: string
        example:  Cosplay Rika
      url:
        type: string
        example: https://es.aliexpress.com/item/Higurashi-no-Naku-Koro-ni-Higurashi-When-They-Cry-Rika-Furude-red-Witch-kimono-Cosplay-Costume/32350149865.html?spm=2114.43010208.4.1.P4V9SM
      prize:
        type: string
        example: 57,68 € 
      description:
        type: string
        example: Higurashi no Naku Koro ni Higurashi When They Cry Rika Furude red Witch kimono Cosplay Costume anysize
  Video:
    required:
      - id
      - title
      - url
    properties:
      id:
        type: string
        example: v1
      title:
        type: string
        example:  Higurashi no Naku Koro ni by Eiko Shimamiya
      url:
        type: string
        example: https://www.youtube.com/watch?v=XgWkIrFQQE0
  
      imagenes:
        $ref: '#/definitions/Imagen'
      noticias:
        $ref: '#/definitions/Noticia'
      productos:
        $ref: '#/definitions/Producto'
      videos:
        $ref: '#/definitions/Video'